pass "uncropped" signal and time to fit func
previous behavior was inconsistent insorfar as `time=None` was passed when fitting linearly while `time=self.time` was passed when fitting in parrallel. time must be passed in order for the baseline to be computed for each trial. furthermore, passing the uncropped signal and time vector is necessary to avoid multiple crops being applied. The update now passes the uncropped version in both instances, fixing the bsaeline issue while avoiding multiple crops. However, the `uncropped_signal` and `uncropped_time` are still not fully integrated into the docstrings and `OBJ_DESC`